Text to Excel Font
The program ask for a string input and converts it into a excel file (run this to see it). (not a great thing though :P )
Limitations till now - It will only convert A-Z.
This program will not behave properly if you give digits, special character, lower case alphabets and even space as a part of the input.